refactor: 修改DateType的length类型为int

+++ src/main/java/org/origami/table/auto/dialect/DateType.java
@@ -1,16 +1,18 @@
 package org.origami.table.auto.dialect;
 
 /**
+ * 数据库数据类型
+ *
  * @author origami
  * @date 2023/8/15 21:47
  */
 public interface DateType {
-    
+
     String getTemplate();
-    
-    Long getLength();
-    
+
+    Integer getLength();
+
     Integer getScale();
-    
+
     Integer getPrecision();
 }
